docs: adjust artemis docs
