+++ docs/modules/artemis.md
@@ -17,7 +17,11 @@ go get github.com/testcontainers/testcontainers-go/modules/artemis
 ## Usage example
 
 <!--codeinclude-->
-[Creating and connecting to an Artemis container](../../modules/artemis/example_test.go) inside_block:ExampleRunContainer
+[Creating to an Artemis container](../../modules/artemis/examples_test.go) inside_block:runArtemisContainer
+<!--/codeinclude-->
+
+<!--codeinclude-->
+[Connecting to an Artemis container](../../modules/artemis/examples_test.go) inside_block:connectToArtemisContainer
 <!--/codeinclude-->
 
 ## Module reference
@@ -58,7 +62,7 @@ If you need to change the default admin credentials (i.e. `artemis:artemis`) use
 [With credentials](../../modules/artemis/artemis_test.go) inside_block:withCredentials
 <!--/codeinclude-->
 
-#### Anonymous Login
+#### Enable Anonymous login
 
 If you need to enable anonymous logins (which are disabled by default) use `WithAnonymousLogin`.
 
@@ -71,7 +75,9 @@ If you need to enable anonymous logins (which are disabled by default) use `With
 If you need to pass custom arguments to the `artemis create` command, use `WithExtraArgs`.
 The default is `--http-host 0.0.0.0 --relax-jolokia`.
 Setting this value will override the default.
-See the documentation on `artemis create` for available options.
+
+!!!info
+    Please see the documentation on `artemis create` for the available options here: [https://activemq.apache.org/components/artemis/documentation/latest/using-server.html#options](https://activemq.apache.org/components/artemis/documentation/latest/using-server.html#options)
 
 <!--codeinclude-->
 [With Extra Arguments](../../modules/artemis/artemis_test.go) inside_block:withExtraArgs
@@ -96,7 +102,7 @@ The Artemis container exposes the following methods:
 User returns the administrator username.
 
 <!--codeinclude-->
-[Retrieving the Administrator User](../../modules/artemis/example_test.go) inside_block:containerUser
+[Retrieving the Administrator User](../../modules/artemis/examples_test.go) inside_block:containerUser
 <!--/codeinclude-->
 
 #### Password
@@ -104,7 +110,7 @@ User returns the administrator username.
 Password returns the administrator password.
 
 <!--codeinclude-->
-[Retrieving the Administrator Password](../../modules/artemis/example_test.go) inside_block:containerPassword
+[Retrieving the Administrator Password](../../modules/artemis/examples_test.go) inside_block:containerPassword
 <!--/codeinclude-->
 
 #### BrokerEndpoint
